Search

+++ Library/ShowData.aspx.cs
@@ -29,7 +29,24 @@ namespace Library
             dt.Columns.Add("Description", typeof(string));
 
             MySqlCommand command = handler.connetion.CreateCommand();
-            command.CommandText = "SELECT * FROM books";
+
+            string Authors = Request.QueryString["tbAuthors"];
+            string Title = Request.QueryString["tbTitle"];
+            string Description = Request.QueryString["tbDescription"];
+            string ReleaseDate = Request.QueryString["tbReleaseDate"];
+            string ISBN = Request.QueryString["tbISBN"];
+            string Format = Request.QueryString["tbFormat"];
+            string Pages = Request.QueryString["tbPages"];
+
+            lbInfo.Text = Authors;
+            command.CommandText = "SELECT * FROM books WHERE " +
+                $"Authors LIKE '{Authors}%' AND " +
+                $"Title LIKE '{Title}%' AND " +
+                $"Description LIKE '{Description}%' AND " +
+                $"Release_date LIKE '{ReleaseDate}%' AND " +
+                $"ISBN LIKE '{ISBN}%' AND " +
+                $"Format LIKE '{Format}%' AND " +
+                $"Pages LIKE '{Pages}%'";
 
             MySqlDataReader reader = command.ExecuteReader();
             string authors, title, releaseDate, isbn, format, description;
@@ -70,34 +87,42 @@ namespace Library
         {
             
         }
-
+        int i = 0;
         protected void GridView1_RowDeleting(object sender, GridViewDeleteEventArgs e)
         {
 
-            GridView gv = (GridView)sender;
-            GridViewRow row = (GridViewRow)gv.Rows[e.RowIndex];
-            int id = Convert.ToInt32(row.Cells[2].Text);
             
-            MySqlCommand command = handler.connetion.CreateCommand();
-            command.CommandText = $"DELETE FROM books WHERE Id='{id}'";
-
-            command.ExecuteNonQuery();
-            
-            
-            gv.DeleteRow(e.RowIndex);
-            Response.Redirect(Request.Url.AbsoluteUri);
         }
 
         protected void GridView1_RowCommand(object sender, GridViewCommandEventArgs e)
         {
-            GridView gv = (GridView)sender;
-            GridViewRow row = (GridViewRow)gv.Rows[Convert.ToInt32(e.CommandArgument)];
-            int id = Convert.ToInt32(row.Cells[2].Text);
+            if (e.CommandName == "EditRow")
+            {
+                GridView gv = (GridView)sender;
+                GridViewRow row = (GridViewRow)gv.Rows[Convert.ToInt32(e.CommandArgument)];
+                int id = Convert.ToInt32(row.Cells[2].Text);
 
-            Session["editId"] = id.ToString();
+                Session["editId"] = id.ToString();
 
-            Response.Redirect("/editRecord.aspx");
+                Response.Redirect("/editRecord.aspx");
+            }
+            else
+            {
+                int rowId = Convert.ToInt32(e.CommandArgument);
+                GridView gv = (GridView)sender;
+                GridViewRow row = (GridViewRow)gv.Rows[rowId];
+                int id = Convert.ToInt32(row.Cells[2].Text);
 
+                MySqlCommand command = handler.connetion.CreateCommand();
+                command.CommandText = $"DELETE FROM books WHERE Id='{id}'";
+                lbInfo.Text = i.ToString();
+                command.ExecuteNonQuery();
+                i++;
+
+                gv.DeleteRow(rowId);
+                Session["editId"] = null;
+                Response.Redirect("/editRecord.aspx");
+            }
         }
     }
 }
